multiInput props interface localized
